+++ mood_music.client/src/api/lastFmApi.ts
@@ -8,3 +8,12 @@ export const getTracksByTags = async () => {
         throw error;
     }
 }
+
+export const getTracksByUserTags = async (tags: string) => {
+    try {
+        const response = await axios.get(`${import.meta.env.VITE_BACKEND_URL}/api/lastfm/tags/${tags}`);
+        return response.data;
+    } catch (error) {
+        throw error;
+    }
+}